test: migrate gas estimates spec to POM
