+++ test/e2e/tests/transaction/gas-estimates.spec.ts
@@ -1,0 +1,331 @@
+import { MockttpServer } from 'mockttp';
+import { loginWithBalanceValidation } from '../../page-objects/flows/login.flow';
+import { createInternalTransaction } from '../../page-objects/flows/transaction';
+import { withFixtures } from '../../helpers';
+import FixtureBuilder from '../../fixtures/fixture-builder';
+import { CHAIN_IDS } from '../../../../shared/constants/network';
+import { GAS_API_BASE_URL } from '../../../../shared/constants/swaps';
+import { mockSpotPrices } from '../tokens/utils/mocks';
+import { Driver } from '../../webdriver/driver';
+import SendTokenConfirmPage from '../../page-objects/pages/send/send-token-confirmation-page';
+
+const PREFERENCES_STATE_MOCK = {
+  preferences: {
+    showFiatInTestnets: true,
+  },
+  // Enables advanced details due to migration 123
+  useNonceField: true,
+};
+
+const SPOT_PRICES_MOCK = {
+  'eip155:1/slip44:60': {
+    price: 1700,
+    marketCap: 382623505141,
+    pricePercentChange1d: 0,
+  },
+};
+
+/**
+ * Mocks Binance-related API requests for BSC network tests.
+ *
+ * @param mockServer - The mock server instance
+ * @returns Array of mock request handlers
+ */
+async function mockBinanceRelatedRequests(
+  mockServer: MockttpServer,
+): Promise<unknown[]> {
+  return [
+    mockServer
+      .forGet('https://bridge.api.cx.metamask.io/networks/56/topAssets')
+      .thenCallback(() => {
+        return {
+          statusCode: 200,
+          json: [
+            {
+              address: '0x0000000000000000000000000000000000000000',
+              symbol: 'BNB',
+            },
+          ],
+        };
+      }),
+    mockServer
+      .forGet(
+        'https://bridge.api.cx.metamask.io/networks/56/aggregatorMetadata',
+      )
+      .thenCallback(() => {
+        return {
+          statusCode: 200,
+          json: {
+            airswapLight: {
+              color: '#2B71FF',
+              title: 'AirSwap',
+              icon: "data:image/svg+xml,%3csvg width='75' height='31' viewBox='0 0 75 31' fill='none' xmlns='http://www.w3.org/2000/svg'%3e %3cpath fill-rule='evenodd' clip-rule='evenodd' d='M31.4038 12.231H30.1152V19.3099H31.4038V12.231Z' fill='%23FDFDFD'/%3e %3cpath fill-rule='evenodd' clip-rule='evenodd'",
+              iconPng: '',
+            },
+          },
+        };
+      }),
+    mockServer
+      .forGet('https://tokens.api.cx.metamask.io/blocklist')
+      .thenCallback(() => {
+        return {
+          statusCode: 200,
+          json: [],
+        };
+      }),
+    mockServer
+      .forGet('https://bridge.api.cx.metamask.io/networks/56/tokens')
+      .thenCallback(() => {
+        return {
+          statusCode: 200,
+          json: [
+            {
+              address: '0x7e8bae727abc245181f7abad0a4445114c0ca987',
+              symbol: '7',
+              decimals: 9,
+              name: 'Lucky7',
+              iconUrl:
+                'https://static.cx.metamask.io/api/v1/tokenIcons/56/0x7e8bae727abc245181f7abad0a4445114c0ca987.png',
+              type: 'erc20',
+              aggregators: ['pancakeCoinGecko'],
+              occurrences: 4,
+              erc20Permit: false,
+              storage: {
+                balance: 5,
+                approval: 3,
+              },
+              blocked: false,
+            },
+          ],
+        };
+      }),
+  ];
+}
+
+/**
+ * Mocks Binance-related requests with the gas API returning an error (422).
+ *
+ * @param mockServer - The mock server instance
+ * @returns Array of mock request handlers
+ */
+async function mockBinanceRelatedRequestsWithGasApiDown(
+  mockServer: MockttpServer,
+): Promise<unknown[]> {
+  const relatedRequests = await mockBinanceRelatedRequests(mockServer);
+  const gasApiDownRequest = mockServer
+    .forGet(
+      `${GAS_API_BASE_URL}/networks/${parseInt(CHAIN_IDS.BSC, 16)}/gasPrices`,
+    )
+    .thenCallback(() => {
+      return {
+        statusCode: 422,
+      };
+    });
+  return [...relatedRequests, gasApiDownRequest];
+}
+
+/**
+ * Mocks Binance-related requests with the gas API available.
+ *
+ * @param mockServer - The mock server instance
+ * @returns Array of mock request handlers
+ */
+async function mockBinanceRelatedRequestsWithGasApiAvailable(
+  mockServer: MockttpServer,
+): Promise<unknown[]> {
+  const relatedRequests = await mockBinanceRelatedRequests(mockServer);
+  const gasApiAvailableRequest = mockServer
+    .forGet(
+      `${GAS_API_BASE_URL}/networks/${parseInt(CHAIN_IDS.BSC, 16)}/gasPrices`,
+    )
+    .thenCallback(() => {
+      return {
+        statusCode: 200,
+      };
+    });
+  return [...relatedRequests, gasApiAvailableRequest];
+}
+
+describe('Gas estimates generated by MetaMask', function () {
+  describe('Send on a network that is EIP-1559 compatible', function () {
+    it('show expected gas defaults', async function () {
+      await withFixtures(
+        {
+          fixtures: new FixtureBuilder()
+            .withPreferencesController(PREFERENCES_STATE_MOCK)
+            .build(),
+          title: this.test?.fullTitle(),
+          testSpecificMock: async (mockServer: MockttpServer) => {
+            await mockSpotPrices(mockServer, SPOT_PRICES_MOCK);
+          },
+        },
+        async ({ driver }: { driver: Driver }) => {
+          await loginWithBalanceValidation(driver);
+
+          await createInternalTransaction(driver);
+
+          const sendTokenConfirmPage = new SendTokenConfirmPage(driver);
+          await sendTokenConfirmPage.checkFirstGasFee('0');
+          await sendTokenConfirmPage.checkNativeCurrency('$0.75');
+        },
+      );
+    });
+
+    it('show expected gas defaults when API is down', async function () {
+      await withFixtures(
+        {
+          fixtures: new FixtureBuilder()
+            .withPreferencesController(PREFERENCES_STATE_MOCK)
+            .build(),
+          testSpecificMock: async (mockServer: MockttpServer) => {
+            mockServer
+              .forGet(`${GAS_API_BASE_URL}/networks/1337/suggestedGasFees`)
+              .thenCallback(() => {
+                return {
+                  json: {
+                    error: 'cannot get gas prices for chain id 1337',
+                  },
+                  statusCode: 503,
+                };
+              });
+            await mockSpotPrices(mockServer, SPOT_PRICES_MOCK);
+          },
+          title: this.test?.fullTitle(),
+        },
+        async ({ driver }: { driver: Driver }) => {
+          await loginWithBalanceValidation(driver);
+
+          await createInternalTransaction(driver);
+
+          const sendTokenConfirmPage = new SendTokenConfirmPage(driver);
+          await sendTokenConfirmPage.checkFirstGasFee('0');
+          await sendTokenConfirmPage.checkNativeCurrency('$0.75');
+        },
+      );
+    });
+
+    it('show expected gas defaults when the network is not supported', async function () {
+      await withFixtures(
+        {
+          fixtures: new FixtureBuilder()
+            .withPreferencesController(PREFERENCES_STATE_MOCK)
+            .build(),
+          testSpecificMock: async (mockServer: MockttpServer) => {
+            mockServer
+              .forGet(`${GAS_API_BASE_URL}/networks/1337/suggestedGasFees`)
+              .thenCallback(() => {
+                return {
+                  statusCode: 422,
+                };
+              });
+            await mockSpotPrices(mockServer, SPOT_PRICES_MOCK);
+          },
+          title: this.test?.fullTitle(),
+        },
+        async ({ driver }: { driver: Driver }) => {
+          await loginWithBalanceValidation(driver);
+
+          await createInternalTransaction(driver);
+
+          const sendTokenConfirmPage = new SendTokenConfirmPage(driver);
+          await sendTokenConfirmPage.checkFirstGasFee('0');
+          await sendTokenConfirmPage.checkNativeCurrency('$0.75');
+        },
+      );
+    });
+  });
+
+  describe('Send on a network that is not EIP-1559 compatible', function () {
+    it('show expected gas defaults on a network supported by legacy gas API', async function () {
+      await withFixtures(
+        {
+          fixtures: new FixtureBuilder()
+            .withPreferencesController(PREFERENCES_STATE_MOCK)
+            .withNetworkControllerOnBnb()
+            .withEnabledNetworks({
+              eip155: {
+                '0x38': true,
+              },
+            })
+            .build(),
+          localNodeOptions: {
+            hardfork: 'berlin',
+            chainId: parseInt(CHAIN_IDS.BSC, 16),
+          },
+          testSpecificMock: async (mockServer: MockttpServer) => {
+            await mockBinanceRelatedRequestsWithGasApiAvailable(mockServer);
+            await mockSpotPrices(mockServer, SPOT_PRICES_MOCK);
+          },
+          title: this.test?.fullTitle(),
+        },
+        async ({ driver }: { driver: Driver }) => {
+          await loginWithBalanceValidation(driver);
+
+          await createInternalTransaction(driver);
+
+          const sendTokenConfirmPage = new SendTokenConfirmPage(driver);
+          await sendTokenConfirmPage.checkFirstGasFee('0');
+          await sendTokenConfirmPage.checkNativeCurrency('$0.01');
+        },
+      );
+    });
+
+    it('show expected gas defaults on a network supported by legacy gas API when that API is down', async function () {
+      await withFixtures(
+        {
+          fixtures: new FixtureBuilder()
+            .withPreferencesController(PREFERENCES_STATE_MOCK)
+            .withNetworkControllerOnBnb()
+            .withEnabledNetworks({
+              eip155: {
+                '0x38': true,
+              },
+            })
+            .build(),
+          localNodeOptions: {
+            hardfork: 'berlin',
+            chainId: parseInt(CHAIN_IDS.BSC, 16),
+          },
+          testSpecificMock: async (mockServer: MockttpServer) => {
+            await mockBinanceRelatedRequestsWithGasApiDown(mockServer);
+            await mockSpotPrices(mockServer, SPOT_PRICES_MOCK);
+          },
+          title: this.test?.fullTitle(),
+        },
+        async ({ driver }: { driver: Driver }) => {
+          await loginWithBalanceValidation(driver);
+
+          await createInternalTransaction(driver);
+
+          const sendTokenConfirmPage = new SendTokenConfirmPage(driver);
+          await sendTokenConfirmPage.checkFirstGasFee('0');
+          await sendTokenConfirmPage.checkNativeCurrency('$0.01');
+        },
+      );
+    });
+
+    it('show expected gas defaults on a network not supported by legacy gas API', async function () {
+      await withFixtures(
+        {
+          fixtures: new FixtureBuilder()
+            .withPreferencesController(PREFERENCES_STATE_MOCK)
+            .build(),
+          localNodeOptions: { hardfork: 'berlin' },
+          title: this.test?.fullTitle(),
+          testSpecificMock: async (mockServer: MockttpServer) => {
+            await mockSpotPrices(mockServer, SPOT_PRICES_MOCK);
+          },
+        },
+        async ({ driver }: { driver: Driver }) => {
+          await loginWithBalanceValidation(driver);
+
+          await createInternalTransaction(driver);
+
+          const sendTokenConfirmPage = new SendTokenConfirmPage(driver);
+          await sendTokenConfirmPage.checkFirstGasFee('0');
+          await sendTokenConfirmPage.checkNativeCurrency('$0.07');
+        },
+      );
+    });
+  });
+});
